run filter on iteration, not on event
Running filter on the event can pass several positive filtered events before they are iterated. So in the case of PB, several things can be queued (and filtered) before anything was collected (making the second filter option incorrect). By moving the filter call, filter will now be in-sync with iterations.